requirements/local.txt
# SHA1:c297e5b8158a6ca20adeb766f65098ac74a0f50e
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r base.txt
alabaster==0.7.12
    # via sphinx
appnope==0.1.3
    # via ipython
astroid==2.12.13
    # via pylint
asttokens==2.2.0
    # via stack-data
attrs==22.1.0
    # via pytest
babel==2.11.0
    # via sphinx
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via furo
black==22.10.0
    # via -r requirements/local.in
bleach==5.0.1
    # via readme-renderer
build==0.9.0
    # via
    #   pip-tools
    #   pytest-checkdocs
bump2version==1.0.1
    # via -r requirements/local.in
commonmark==0.9.1
    # via rich
coverage[toml]==6.5.0
    # via pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
dill==0.3.6
    # via pylint
docutils==0.19
    # via
    #   pytest-checkdocs
    #   readme-renderer
    #   sphinx
exceptiongroup==1.0.4
    # via pytest
executing==1.2.0
    # via stack-data
flake8==6.0.0
    # via
    #   -r requirements/local.in
    #   flake8-pyproject
flake8-pyproject==1.2.1
    # via -r requirements/local.in
furo==2022.9.29
    # via -r requirements/local.in
imagesize==1.4.1
    # via sphinx
importlib-metadata==5.1.0
    # via
    #   keyring
    #   pytest-checkdocs
    #   sphinx
    #   twine
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements/local.in
ipython==8.7.0
    # via
    #   -r requirements/local.in
    #   ipdb
isort==5.10.1
    # via
    #   -r requirements/local.in
    #   pylint
jaraco-classes==3.2.3
    # via keyring
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via sphinx
keyring==23.11.0
    # via twine
lazy-object-proxy==1.8.0
    # via astroid
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
more-itertools==9.0.0
    # via jaraco-classes
mypy==0.991
    # via -r requirements/local.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   build
    #   pytest
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.10.2
    # via black
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-compile-multi==2.6.1
    # via -r requirements/local.in
pip-tools==6.11.0
    # via pip-compile-multi
pkginfo==1.9.2
    # via twine
platformdirs==2.5.4
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.33
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pygments==2.13.0
    # via
    #   furo
    #   ipython
    #   readme-renderer
    #   rich
    #   sphinx
pylint==2.15.7
    # via -r requirements/local.in
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r requirements/local.in
    #   pytest-cov
pytest-checkdocs==2.9.0
    # via -r requirements/local.in
pytest-cov==4.0.0
    # via -r requirements/local.in
python-dateutil==2.8.2
    # via rst-linker
pytz==2022.6
    # via babel
readme-renderer==37.3
    # via twine
requests-toolbelt==0.10.1
    # via twine
rfc3986==2.0.0
    # via twine
rich==12.6.0
    # via twine
rst-linker==2.3.1
    # via -r requirements/local.in
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.3.2.post1
    # via beautifulsoup4
sphinx==5.3.0
    # via
    #   -r requirements/local.in
    #   furo
    #   sphinx-basic-ng
sphinx-basic-ng==1.0.0b1
    # via furo
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
stack-data==0.6.2
    # via ipython
toml==0.10.2
    # via ipdb
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   flake8-pyproject
    #   mypy
    #   pep517
    #   pylint
    #   pytest
tomlkit==0.11.6
    # via pylint
toposort==1.7
    # via pip-compile-multi
traitlets==5.6.0
    # via
    #   ipython
    #   matplotlib-inline
twine==4.0.2
    # via -r requirements/local.in
types-requests==2.28.11.5
    # via -r requirements/local.in
types-urllib3==1.26.25.4
    # via types-requests
typing-extensions==4.4.0
    # via
    #   astroid
    #   black
    #   mypy
    #   pylint
    #   rich
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
wheel==0.38.4
    # via pip-tools
wrapt==1.14.1
    # via astroid
zipp==3.11.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
pip==22.3.1
    # via pip-tools
setuptools==65.6.3
    # via
    #   ipdb
    #   pip-tools
